memset(), memcpy() 함수 파라미터 ulong 에서 long으로 변경

+++ BigImageViewer/FormMain.cs
@@ -59,7 +59,7 @@ namespace BigImageViewer {
             imgBW = (frmW - fwdOvlp) * fwdNum + fwdOvlp;
             imgBH = frmH * frmNum;
             imgBuf = Marshal.AllocHGlobal((IntPtr)((long)imgBW * imgBH));
-            Msvcrt.memset(imgBuf, 0, (ulong)((long)imgBW * imgBH));
+            Msvcrt.memset(imgBuf, 0, (long)imgBW * imgBH);
         }
 
         // 포워드 이미지 로드
@@ -77,7 +77,7 @@ namespace BigImageViewer {
                 if (r) {
                     succNum++;
                 } else {
-                    Msvcrt.memset(buf, 0, (ulong)frmSize);
+                    Msvcrt.memset(buf, 0, frmSize);
                 }
             }
             Log($"Fwd_{ifwd} 이미지 {frmNum}개 중 {succNum}개 이미지 로드 성공");
